pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "reso"
version = "0.1.0"
description = "AI-powered wellness assistant — Zone 1: Ingress & Orchestration"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sse-starlette>=2.1.0",

    # LangGraph / LangChain
    "langgraph>=0.2.70",
    "langgraph-checkpoint-postgres>=2.0.0",
    "psycopg[binary,pool]>=3.1.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",

    # Supabase
    "supabase>=2.10.0",

    # Redis
    "redis[hiredis]>=5.2.0",

    # Auth
    "python-jose[cryptography]>=3.3.0",

    # Config
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",

    # HTTP client
    "httpx>=0.28.0",

    # Observability
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.28.0",

    # Utilities
    "tenacity>=9.0.0",
    "structlog>=24.4.0",

    # Long-term memory with graph support
    "mem0ai[graph]>=0.1.44",
    "qdrant-client>=1.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.32.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "live: tests that make real network calls (requires OPENAI_API_KEY)",
]
